+++ src-tauri/tauri.conf.json
@@ -13,20 +13,21 @@
     "windows": [
       {
         "title": "holdem",
-        "width": 300,
-        "height": 300,
+        "width": 165,
+        "height": 175,
         "dragDropEnabled": true,
         "decorations": false,
         "transparent": true,
         "focus": false,
         "alwaysOnTop": true,
         "skipTaskbar": true,
+        "resizable": false,
         "theme": "Dark"
       },
       {
         "title": "popup",
-        "width": 300,
-        "height": 300,
+        "width": 350,
+        "height": 280,
         "create": false,
         "transparent": true,
         "decorations": false
